Rename the module containing exporting logic to avoid confusion
